Introduce subcommands for main binary, add scanner functionality
- main binary now allows only two commands: webhook, scanner
- "webhook" command starts validation webhook same way as before
- "scanner" allows to do one-time outside of cluster scan for objects that are violating validation rules
